+++ CHANGELOG.md
@@ -5,7 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.4] - Latest
+## [0.1.6] - Latest
+
+### Added
+
+- A first iteration of Funnel custom dimensions and metrics resources.
+- A data source for workspaces
+
+### Changed
+
+- Updated the terraform-plugin-framework to v1.19.0.
+
+## [0.1.5] 
+
+### Changed
+
+- Switch the Auth0 URLs.
+
+### Added
+
+- Import functionality for the Workspace resource.
+
+## [0.1.4]
 
 ### Added
 
@@ -15,7 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The old workspace data source.
 
-## [0.1.3] - Latest
+## [0.1.3]
 
 ### Added